Adding ability to populate annual balance sheet.

+++ pyvalue/data/balance_sheet.py
@@ -3,9 +3,9 @@ A module that represents a balance sheet of a stock as of a specific date.
 Author: Emre Tezel
 """
 
-from sqlalchemy import Column, Integer, ForeignKey, Date
+from sqlalchemy import Column, Integer, ForeignKey, Date, Float
 from sqlalchemy.orm import relationship
-from common import Base
+from pyvalue.data import Base
 
 
 class BalanceSheet(Base):
@@ -14,7 +14,7 @@ class BalanceSheet(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     stock_id = Column(Integer, ForeignKey("stocks.id"), nullable=False)
     date = Column(Date, nullable=False)
-    total_assets = Column(Integer, nullable=False)
-    total_liabilities = Column(Integer, nullable=False)
+    total_assets = Column(Float, nullable=False)
+    total_liabilities = Column(Float, nullable=False)
 
     stock = relationship("Stock", back_populates="balance_sheets")
